state saving feature added
delete game feature added

major bug for skipping player's kala fixed

+++ src/main/java/com/codebase/codechallenge/kalagame/repository/KalaGameRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 @org.springframework.stereotype.Repository
-public interface KalaGameRepository extends JpaRepository<GameEntity,Long> {
+public interface KalaGameRepository extends JpaRepository<GameEntity,Integer> {
 
 }
